add 14.0b6, 13 components reports, and don't save ADU counts when we don't have any

+++ get-explosives.php
@@ -340,7 +340,8 @@ foreach ($reports as $rep) {
 
     // Save ADUs (if needed).
     $anafadu = $anadir.'/'.$fadu;
-    if (!$rep['fake_adu'] && (!file_exists($anafadu) || !filesize($anafadu))) {
+    if (intval(@$adu[$anadir]) && !$rep['fake_adu'] &&
+        (!file_exists($anafadu) || !filesize($anafadu))) {
       print('Saving ADU count'."\n");
       file_put_contents($anafadu, $adu[$anadir]);
     }
